[NGC-4548] Updated to Play 2.6
I spent a lot of time trying to figure out how to create an instance of `MessagesControllerComponents`, required by `AuthStubController`. The result is in `AuthStubControllerSpec` and is as small as I could boil it down to. Since all controllers that extend `FrontendController` for Play 2.6 are going to need this I suggest we extract it to a library (ideally `hmrctest`) so that other projects can avoid the boilerplate.